alx_travel_app/requirement.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
drf-yasg
celery
rabbitmq
django-environ
django-seed
psycopg2